requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
dotenv
strands-agents[ollama]
chromadb
langchain-community
langchain-text-splitters
sentence-transformers
pymupdf
pypdf